Resolve conflicts preserving shielded pool features from HEAD while
adopting v1.0-dev improvements: refactored asset lock fee calculation,
mine dialog hardening, network validation, and API signature updates.

+++ src/backend_task/mod.rs
@@ -261,7 +261,7 @@ pub enum BackendTaskSuccessResult {
         recovered_count: usize,
         total_amount: u64,
     },
-    MineBlocksSuccess(usize),
+    MineBlocksSuccess(u64),
 
     // DPNS operation results (replacing string messages)
     ScheduledVotes,
